MCPLangchain/requirements.txt
mcp
langchain
langgraph
langchain-groq
python-dotenv
httpx
uvicorn
beautifulsoup4